Nothing printed when specifying a very large maximum layout width https://bugs.webkit.org/show_bug.cgi?id=64831
Reviewed by Beth Dakin.

I am not adding a test because the DumpRenderTree printing test machinery is based on PrintContext,
which does not exercise the code path on which this bug lies.

* page/FrameView.cpp:
(WebCore::FrameView::forceLayoutForPagination): Fixed an integer overflow.



Canonical link: https://commits.webkit.org/80557@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@91314 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/page/FrameView.cpp
@@ -2632,7 +2632,7 @@ void FrameView::forceLayoutForPagination(const FloatSize& pageSize, float maximu
         bool horizontalWritingMode = root->style()->isHorizontalWritingMode();
         int docLogicalWidth = horizontalWritingMode ? root->documentRect().width() : root->documentRect().height();
         if (docLogicalWidth > pageLogicalWidth) {
-            flooredPageLogicalWidth = std::min<int>(docLogicalWidth, pageLogicalWidth * maximumShrinkFactor);
+            flooredPageLogicalWidth = std::min<float>(docLogicalWidth, pageLogicalWidth * maximumShrinkFactor);
             if (pageLogicalHeight)
                 root->setPageLogicalHeight(flooredPageLogicalWidth / pageSize.width() * pageSize.height());
             root->setLogicalWidth(flooredPageLogicalWidth);
